修复DAC日志中新添加的uint8类型的字段不能正常显示的问题
git-svn-id: http://svn01/repositories/r01/p2p/trunk@1660 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ p2sp/download/LiveDownloadDriver.cpp
@@ -753,7 +753,7 @@ namespace p2sp
         log_stream << "&G1=" << info.MaxUploadSpeedIncludeSameSubnet;
         log_stream << "&H1=" << info.MaxUploadSpeedExcludeSameSubnet;
         log_stream << "&I1=" << info.MaxUnlimitedUploadSpeedInRecord;
-        log_stream << "&J1=" << info.ChangeToP2PConditionWhenStart;
+        log_stream << "&J1=" << (uint32_t)info.ChangeToP2PConditionWhenStart;
         log_stream << "&K1=" << info.ChangedToHttpTimesWhenUrgent;
         log_stream << "&L1=" << info.BlockTimesWhenUseHttpUnderUrgentSituation;
